Removes unnecessary dependency on ROS master node (#18)
The rospy.init_node method was being called (for creating
timestamps with now's time), but using the python time module for
seeding the object is sufficient (wall-clock).

Removes the rospy.init_node calls, which turns the whole test class
into a UNIT test WITHOUT need for a ROS master node.