Decode base64-encoded byte arrays as unicode
This changes the behavior for binary array message fields (uint8[],
char[] etc.) that are converted to dictionaries as follows:

* in python2: type "str"   -> type "unicode"
* in python3: type "bytes" -> type "str"

This should be okay, because the result of calling standard_b64decode on
those fields doesn't change.

This change is required for python3 compatibility, because the "bytes"
type cannot be serialized to json. This broke the following tests:

* test_json_message_converter.test_ros_message_with_3uint8_array
* test_json_message_converter.test_json_with_string